requirements.txt
allure_python_commons==2.13.1
pytest==7.3.0
